requirements.txt
opencv-python>=4.5.5
mediapipe>=0.10.0
pyserial>=3.5
numpy>=1.21.0
Pillow>=9.0.0
psutil>=5.8.0
